requirements.txt
# Core ML/DL Libraries
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0  # PyTorch Image Models (for EfficientNet, ConvNeXt, etc.)
onnx>=1.14.0
onnxruntime>=1.15.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
opencv-python>=4.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Training Utilities
tqdm>=4.65.0
tensorboard>=2.13.0
scikit-learn>=1.3.0

# Azure SDK
azure-ai-ml>=1.11.0
azure-identity>=1.14.0
azureml-core>=1.52.0

# Kaggle
kaggle>=1.5.16

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
